+++ packages/benchmarks/src/envVars.ts
@@ -1,12 +1,4 @@
-export const BRAINTRUST_ENV_VARS = {
-  BRAINTRUST_API_KEY: "",
-  BRAINTRUST_ENDPOINT: "",
-};
-
-export const GCP_VERTEX_AI_ENV_VARS = {
-  GCP_OPENAI_ENDPOINT: "",
-  GCP_API_KEY: "",
-};
+import { GCP_VERTEX_AI_ENV_VARS } from "mongodb-rag-core";
 
 export const envVars = {
   MONGODB_DATABASE_NAME: "",
