+++ packages/mongodb-rag-core/.env.example
@@ -11,3 +11,7 @@ MONGODB_CONNECTION_URI=<connection_uri>
 MONGODB_DATABASE_NAME=<docs-chatbot-ENV>
 VECTOR_SEARCH_INDEX_NAME=<index_name>
 OPENAI_API_VERSION="2024-06-01"
+BRAINTRUST_ENDPOINT="https://api.braintrust.dev/v1/proxy"
+BRAINTRUST_API_KEY="<some api key>"
+GCP_OPENAI_ENDPOINT="https://{LOCATION}-aiplatform.googleapis.com/v1beta1/projects/{PROJECT_ID}/locations/{LOCATION}/endpoints/openapi"
+GCP_API_KEY="SOME_API_KEY"
