+++ packages/mongodb-rag-core/package.json
@@ -32,6 +32,7 @@
     "./openai": "./build/openai.js",
     "./braintrust": "./build/braintrust.js",
     "./dataSources": "./build/dataSources/index.js",
+    "./models": "./build/models/index.js",
     "./eval": "./build/eval/index.js"
   },
   "scripts": {
