+++ packages/mongodb-rag-core/src/CoreEnvVars.ts
@@ -36,3 +36,13 @@ export const CORE_ENV_VARS = {
   ...CORE_CHATBOT_APP_ENV_VARS,
   VECTOR_SEARCH_INDEX_NAME: "",
 };
+
+export const BRAINTRUST_ENV_VARS = {
+  BRAINTRUST_API_KEY: "",
+  BRAINTRUST_ENDPOINT: "",
+};
+
+export const GCP_VERTEX_AI_ENV_VARS = {
+  GCP_OPENAI_ENDPOINT: "",
+  GCP_API_KEY: "",
+};
